[fix] update comment entity
